Move around and fix up types
